requirements.txt
langchain==0.0.212
google-cloud-aiplatform==1.26.1
google==3.0.0
streamlit-nightly==1.23.2.dev20230622